Made improvements to file loading/handling.
Files can now be deleted through the load level window.
Bug where the file list wouldn't be updated until the scene was reloaded
if a new level was saved was fixed.
Improved the way file entry prefabs were positioned in the file window.
Level name is now cleared when the clear level button is pressed in the
editor panel.